SH-6 Refactored code according to comments
