+++ src/app/app.module.ts
@@ -1,11 +1,8 @@
 import { isDevMode, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { SharedModule } from 'src/shared/shared.module';
-import { ProductsModule } from './pages/products/products.module';
-
 import { AppComponent } from './app.component';
 import { RouterModule, Routes } from '@angular/router';
-import { ProductDashboardComponent, ProductPageComponent } from './pages/products/components';
 
 import { Route } from '@shared-module';
 import { StoreModule } from '@ngrx/store';
@@ -18,8 +15,14 @@ import { TranslateModule } from '@ngx-translate/core';
 
 const routes: Routes = [
   { path: Route.base, redirectTo: Route.products, pathMatch: 'full' },
-  { path: Route.products, component: ProductDashboardComponent },
-  { path: Route.productById, component: ProductPageComponent },
+  {
+    path: Route.products,
+    loadChildren: () => import('./pages/products/products.module').then(({ ProductsModule }) => ProductsModule),
+  },
+  {
+    path: Route.productById,
+    loadChildren: () => import('./pages/products/products.module').then(({ ProductsModule }) => ProductsModule),
+  },
   {
     path: Route.cart,
     loadChildren: () => import('./pages/cart/cart.module').then(({ CartModule }) => CartModule),
@@ -37,7 +40,6 @@ const routes: Routes = [
 @NgModule({
   imports: [
     BrowserModule,
-    ProductsModule,
     RouterModule.forRoot(routes),
     SharedModule,
     StoreModule.forRoot({}),
