--- src/app/pages/products/services/filter/index.ts
@@ -1,3 +1,0 @@
-export * from './filter.service';
-export * from './filter-facade.service';
-export * from './sort-facade.service';
