+++ src/app/pages/products/services/index.ts
@@ -1,2 +1,1 @@
 export * from './card-state.service';
-export * from './filter';
