+++ src/app/pages/products/store/index.ts
@@ -1,1 +1,1 @@
-export * from './filters';
+
