+++ src/shared/services/filter-facade.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { filter, Observable } from 'rxjs';
 import { ICalculatedProduct, IFilterDefinition } from '@shared-module';
 import { Store } from '@ngrx/store';
-import { FilterActions, FilterSelectors } from '../../store';
+import { FilterActions, FilterSelectors } from '../store/filters';
 
 @Injectable()
 export class FilterFacadeService {
@@ -16,10 +16,6 @@ export class FilterFacadeService {
   getFilterValue(): Observable<Record<string, string[]>> {
     return this.store.select(FilterSelectors.selectFilter)
       .pipe(filter(Boolean));
-  }
-
-  setFilterValue(selectedFilters: Record<string, string[]>): void {
-    this.store.dispatch(FilterActions.setSelectedFilters({ selectedFilters: selectedFilters }));
   }
 
   getManipulatedProducts(): Observable<ICalculatedProduct[]> {
