+++ src/shared/services/index.ts
@@ -6,3 +6,6 @@ export * from './category';
 export * from './cart';
 export * from './toast.service';
 export * from './orders';
+export * from './filter.service';
+export * from './filter-facade.service';
+export * from './sort-facade.service';
