+++ src/shared/store/filters/index.ts
@@ -1,3 +1,2 @@
 export * from './filter.actions';
 export * from './filter.selectors';
-export * from './filter.effect';
